requirements.txt
pandas==3.0.2
pyarrow==23.0.1
plotly==6.7.0
groq==1.1.2
streamlit==1.56.0
python-dotenv==1.2.2
